requirements.txt
affine==2.4.0
albumentations==1.3.1
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.5
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
filelock==3.9.0
fonttools==4.41.0
idna==3.4
imageio==2.31.1
importlib-resources==6.0.0
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
lazy_loader==0.3
lightning-utilities==0.9.0
MarkupSafe==2.1.2
matplotlib==3.7.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
opencv-python-headless==4.8.0.74
packaging==23.1
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
rasterio==1.3.8
requests==2.28.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
shutup==0.2.0
six==1.16.0
snuggs==1.4.7
sympy==1.11.1
threadpoolctl==3.2.0
tifffile==2023.7.10
torch~=2.0.1
torchaudio==2.0.2+cu117
torchmetrics==1.0.1
torchvision==0.15.2+cu117
tqdm==4.65.0
typing_extensions==4.4.0
urllib3==1.26.13
zipp==3.16.2

ray~=2.6.1
opencv-python~=4.8.0.74
imgaug~=0.4.0